Remove cover element of SearchNav when it is hidden
The original way to show/hide the cover element is not work well, user can not click anything. The reason is that once the z-index of this cover element parent is higher than the others, the cover element is always in the top layer even it is hidden, so remove the element when it is hidden is the best way to prevent this situation.